feat: add mock for auth (currentUser, login, logout)
